Merging pull request #13.
A thank you again to @tomvandenberge for that pull request. I was
able to integrate it easily into the reporter-config2 package
(support for metrics 2.x). There is a design problem in metrics 3.x.
All of the reporters have a builder that accepts a Registry but they
forgot to add to their parent interface a method that accepts a Registry
and generates a builder. They will also need to add an interface for
the Builder. I'll submit a pull request to metrics to clean that up
and then we can adopt this feature to the reporter-config3 package.

I will continue to clean up this pull request to make it suitable for landing.